editCartItem route created, deleteCartItem.js renamed due to typo
