+++ source/routes/cart.js
@@ -1,8 +1,10 @@
 const router = require("express").Router();
 const postCartItem = require("../controllers/Cart/postCartItem");
-const deleteCartItem = require("../controllers/Cart/delelteCartItem");
+const deleteCartItem = require("../controllers/Cart/deleteCartItem");
+const editCartItem = require("../controllers/Cart/editCartItem");
 
 router.post("/", postCartItem);
 router.delete("/", deleteCartItem);
+router.put("/", editCartItem);
 
 module.exports = router;
